requirements.txt
requests
python-dotenv
google-generativeai
flask
flask-cors